backend/requirements.txt
# ==========================================
# NYAYAK - Indian Legal AI Assistant
# Backend Dependencies
# ==========================================

# WEB FRAMEWORK
fastapi>=0.115.0,<0.116
uvicorn[standard]>=0.32.0,<0.33
python-multipart>=0.0.22,<0.0.25
pydantic>=2.5.0,<3.0

# LANGUAGE MODEL & NLP
transformers>=4.40.0,<4.46
torch>=2.2.0,<2.4
sentence-transformers>=2.7.0,<3.0
# Only one entry for FAISS (stable Linux version)
faiss-cpu==1.9.0.post1

# RETRIEVAL AUGMENTED GENERATION (RAG)
langchain==1.2.10
langchain-core==1.2.13
langchain-community==0.4.1
langchain-huggingface==1.2.0

# UTILITIES
numpy>=1.24.0,<2.0
pypdf>=4.0.0,<5.0
google-genai
python-dotenv
